requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
sqlalchemy==2.0.36
pydantic==2.10.2
pyjwt==2.10.0

